requirements.txt
Pillow>=10.0.0
img2pdf>=0.4.4
python-telegram-bot[job-queue]>=20.0
python-dotenv>=1.0.0
